LayoutTests/media/modern-media-controls/skip-back-button/skip-back-button.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>SkipBackButton</code> class.");

const skipBackButton = new SkipBackButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("skipBackButton.element.localName", "button");
shouldBeTrue('skipBackButton.element.classList.contains("icon")');
shouldBeTrue('skipBackButton.element.classList.contains("skip-back")');
shouldBeEqualToString("skipBackButton.iconName", `${Icons.SkipBack}`);

const iOSSkipBackButton = new SkipBackButton({ layoutTraits: LayoutTraits.iOS });

shouldBecomeEqual('skipBackButton.element.style.webkitMaskImage.includes("macOS/interval-skip-back@")', "true", () => {
    shouldBecomeEqual('iOSSkipBackButton.element.style.webkitMaskImage.includes("iOS/interval-skip-back@")', "true", finishJSTest);
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
